Create Scraping_tool_code_base
This file may be a bit out of date but last I checked it was working with my Arduino environment. I am on version 1.8.8. It was also written using a mac which may or may not cause problems. The code is very simple, thresholds can be altered for demonstration purposes.