Update Physical_Therapy_Tool_code_-base_V2.ino
This is a later version of the code. It ran for the first user study. Due to the lack of good control of code, this may be the latest version and most usable. Thresholds can be altered as needed but these were set for the specific force applied by the Physical Therapists.